Finishing up basic service
A request with an address will now return the list of Registered Community Organizations for that address should it be available.  There is basic error handling, though those should be converted to specific error response codes.

In .env, if APP_ENV is production, the application will reach out to the external services.  If it is anything else (local, develop, etc), they are Mocked.