I have created and outline for the web app to be built on. The pages as well as the basic components that build these pages have been created. On top of this I have also set up the router so that we can now link to different pages via routerLink='/'. We will not be using href unless we are going outside the angular2 app. All routing calls can be found at src/app/app.routing.ts. In this document you will see the path names.
One thing also to note is this is just a very rough outline. All feedback is welcome. We may also need to tighten up the naming con. Also, if anyone has input on this please let me know. When I changed the pages export name from 'AboutComponent' to 'AboutPage' I noticed the build time for the page extended. I do not know if this will flow over to a live environment or not.

To Do: The services still need to be created so we can import the api information and start populating it within the app.