Add all route
We want a route that can be used to pull in every RCO that's available to us, so we can store them and start doing things with them.  This route does that.

Ideally we have an event listener for Organization creation and use that to make tokens and send out emails.  For now, just make the tokens manually.