Changing this service
We VERY rarely want to be hitting the actual remote GIS service because it is slow.  Only hit it once per day with the getAll call, which will update all of our local Polygon information.  From then on, we can run our own polygon searches which are _much, much_ faster.

Adds Kernel job to run the getAll function once a day to keep that fairly up to date.